refactor: move files to feature directories
