+++ src/__tests__/page-header/title-slice.test.ts
@@ -1,5 +1,5 @@
 import { AnyAction } from 'redux'
-import title, { updateTitle, changeTitle } from '../../slices/title-slice'
+import title, { updateTitle, changeTitle } from '../../page-header/title-slice'
 
 describe('title slice', () => {
   describe('title reducer', () => {
