+++ src/__tests__/patients/new/NewPatientForm.test.tsx
@@ -10,7 +10,7 @@ import SelectWithLabelFormGroup from '../../../components/input/SelectWithLableF
 import DatePickerWithLabelFormGroup from '../../../components/input/DatePickerWithLabelFormGroup'
 import TextFieldWithLabelFormGroup from '../../../components/input/TextFieldWithLabelFormGroup'
 import Patient from '../../../model/Patient'
-import { getPatientName } from '../../../util/patient-name-util'
+import { getPatientName } from '../../../patients/util/patient-name-util'
 
 const onSave = jest.fn()
 const onCancel = jest.fn()
