+++ src/dashboard/Dashboard.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useTranslation } from 'react-i18next'
-import useTitle from '../util/useTitle'
+import useTitle from '../page-header/useTitle'
 
 const Dashboard: React.FC = () => {
   const { t } = useTranslation()
