+++ src/page-header/useTitle.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from 'react'
 import { useDispatch } from 'react-redux'
-import { updateTitle } from '../slices/title-slice'
+import { updateTitle } from './title-slice'
 
 export default function useTitle(title: string): void {
   const dispatch = useDispatch()
