+++ src/patients/view/ViewPatient.tsx
@@ -4,14 +4,14 @@ import { useParams, withRouter } from 'react-router-dom'
 import { Panel, Spinner } from '@hospitalrun/components'
 import { useTranslation } from 'react-i18next'
 import { differenceInYears } from 'date-fns'
-import useTitle from '../../util/useTitle'
+import useTitle from '../../page-header/useTitle'
 import { fetchPatient } from '../patient-slice'
 import { RootState } from '../../store'
 import TextFieldWithLabelFormGroup from '../../components/input/TextFieldWithLabelFormGroup'
 import TextInputWithLabelFormGroup from '../../components/input/TextInputWithLabelFormGroup'
 import SelectWithLabelFormGroup from '../../components/input/SelectWithLableFormGroup'
 import DatePickerWithLabelFormGroup from '../../components/input/DatePickerWithLabelFormGroup'
-import { getPatientFullName } from '../../util/patient-name-util'
+import { getPatientFullName } from '../util/patient-name-util'
 import Patient from '../../model/Patient'
 
 const getPatientAge = (dateOfBirth: string | undefined): string => {
