+++ src/store/index.ts
@@ -2,8 +2,8 @@ import { configureStore, combineReducers, Action } from '@reduxjs/toolkit'
 import ReduxThunk, { ThunkAction } from 'redux-thunk'
 import patient from '../patients/patient-slice'
 import patients from '../patients/patients-slice'
-import title from '../slices/title-slice'
-import user from '../slices/user-slice'
+import title from '../page-header/title-slice'
+import user from '../user/user-slice'
 
 const reducer = combineReducers({
   patient,
